Moved the file API/BusinessModel.js -> models/business.js
