+++ true-business-backend/controllers/userController.js
@@ -1,5 +1,4 @@
 const bcrypt = require('bcrypt');
-const mongoose = require('mongoose');
 const User = require('../models/user');
 const bcryptRounds = 10;
